base/api/serializers.py
#el serialize toma un model de dato y lo transforma en un modelo JSON

from rest_framework.serializers import ModelSerializer
from base.models import Room

class RoomSerializer(ModelSerializer):
    class Meta:
        model=Room
        fields='__all__'