requirements.txt
asgiref==3.4.1
distlib==0.3.4
dj-database-url==1.0.0
Django==3.2.9
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.15.1
whitenoise==6.2.0
